misc: remove old to_request and revamp to_params
The new ToParams trait does not force us to allocate storage upfront
and allows the caller to use their own as long as it is Extend'able.
